requirements.txt
azure-core==1.26.3
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
databricks-api==0.8.0
databricks-cli==0.17.5
delta==0.4.2
delta-spark==2.2.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
py4j==0.10.9.5
pycparser==2.21
PyJWT==2.6.0
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
shelljob==0.6.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
